requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
Pillow~=10.1.0
drf_yasg